zheng-driver-hsqldb/src/main/java/com/dadazhishi/zheng/datasource/HsqldbModule.java
package com.dadazhishi.zheng.datasource;

import com.google.inject.Provides;
import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.SQLException;

public class HsqldbModule extends AbstractDataSourceModule {

  public static final String DRIVER_CLASS_NAME = "org.hsqldb.jdbc.JDBCDriver";
  private DataSourceConfig config;

  public HsqldbModule() {
    try {
      Class.forName(DRIVER_CLASS_NAME);
    } catch (ClassNotFoundException e) {
      throw new RuntimeException(e);
    }
  }

  @Override
  public void init(DataSourceConfig config) {
    this.config = config;
  }

  @Provides
  public Connection connection() {
    try {
      return DriverManager
          .getConnection(config.getJdbcUrl(), config.getUsername(), config.getPassword());
    } catch (SQLException e) {
      throw new RuntimeException(e);
    }
  }
}
